TSK-581 Separate configuration of LdapTemplate. Wire LdapTemplate conditionally.
