+++ rest/taskana-rest-spring/src/main/java/pro/taskana/rest/RestConfiguration.java
@@ -4,18 +4,14 @@ import java.sql.SQLException;
 
 import javax.sql.DataSource;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Scope;
-import org.springframework.core.env.Environment;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
 import org.springframework.http.converter.json.SpringHandlerInstantiator;
-import org.springframework.ldap.core.LdapTemplate;
-import org.springframework.ldap.core.support.LdapContextSource;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 import com.fasterxml.jackson.databind.cfg.HandlerInstantiator;
@@ -36,44 +32,6 @@ import pro.taskana.ldap.LdapClient;
 @ComponentScan
 @EnableTransactionManagement
 public class RestConfiguration {
-
-    @Autowired
-    private Environment env;
-
-    @Bean
-    public LdapContextSource contextSource() {
-
-        LdapContextSource contextSource = new LdapContextSource();
-        boolean useLdap;
-        String useLdapConfigValue = env.getProperty("taskana.ldap.useLdap");
-        if (useLdapConfigValue == null || useLdapConfigValue.isEmpty()) {
-            useLdap = false;
-        } else {
-            useLdap = Boolean.parseBoolean(useLdapConfigValue);
-        }
-        if (useLdap) {
-            contextSource.setUrl(env.getRequiredProperty("taskana.ldap.serverUrl"));
-            contextSource.setBase(env.getRequiredProperty("taskana.ldap.baseDn"));
-            contextSource.setUserDn(env.getRequiredProperty("taskana.ldap.bindDn"));
-            contextSource.setPassword(env.getRequiredProperty("taskana.ldap.bindPassword"));
-        } else {
-            contextSource.setUrl("ldap://localhost:9999");
-            contextSource.setBase("o=taskana");
-            contextSource.setUserDn("user");
-            contextSource.setPassword("secret");
-        }
-        return contextSource;
-    }
-
-    @Bean
-    public LdapClient ldapClient() {
-        return new LdapClient();
-    }
-
-    @Bean
-    public LdapTemplate ldapTemplate() {
-        return new LdapTemplate(contextSource());
-    }
 
     @Bean
     public ClassificationService getClassificationService(TaskanaEngine taskanaEngine) {
@@ -105,6 +63,11 @@ public class RestConfiguration {
     @Scope(ConfigurableBeanFactory.SCOPE_SINGLETON)
     public TaskanaEngineConfiguration taskanaEngineConfiguration(DataSource dataSource) throws SQLException {
         return new SpringTaskanaEngineConfiguration(dataSource, true, true);
+    }
+
+    @Bean
+    public LdapClient ldapClient() {
+        return new LdapClient();
     }
 
     // Needed to override JSON De-/Serializer in Jackson.
