+++ rest/taskana-rest-spring/src/test/java/pro/taskana/rest/TestConfiguration.java
@@ -6,11 +6,12 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 
 import pro.taskana.configuration.TaskanaEngineConfiguration;
+import pro.taskana.ldap.LdapConfiguration;
 
 /**
  * Configuration class for all rest tests.
  */
-@Import(RestConfiguration.class)
+@Import({RestConfiguration.class, LdapConfiguration.class})
 public class TestConfiguration {
 
     @Bean
